Extract Feeds::CheckMediumReplyItem service
